requirements.txt
annotated-types==0.7.0
anyio==4.4.0
audioread==3.0.1
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
decorator==5.1.1
distro==1.9.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
exceptiongroup==1.2.2
ffmpeg-downloader==0.3.0
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
kiwipiepy==0.17.1
kiwipiepy_model==0.17.0
konlpy==0.6.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.35.14
openai-whisper==20231117
packaging==24.1
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
tbb==2021.13.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
